+++ hexstrike_mcp.py
@@ -259,798 +259,14 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
     from mcp_tools.binary_analysis_and_reverse_engineering import register_binary_analysis_and_reverse_engineering_tools
     register_binary_analysis_and_reverse_engineering_tools(mcp, hexstrike_client, logger)
 
-    # ============================================================================
-    # ENHANCED BINARY ANALYSIS AND EXPLOITATION FRAMEWORK (v6.0)
-    # ============================================================================
-
-    @mcp.tool()
-    def ghidra_analysis(binary: str, project_name: str = "hexstrike_analysis",
-                       script_file: str = "", analysis_timeout: int = 300,
-                       output_format: str = "xml", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Ghidra for advanced binary analysis and reverse engineering.
-
-        Args:
-            binary: Path to the binary file
-            project_name: Ghidra project name
-            script_file: Custom Ghidra script to run
-            analysis_timeout: Analysis timeout in seconds
-            output_format: Output format (xml, json)
-            additional_args: Additional Ghidra arguments
-
-        Returns:
-            Advanced binary analysis results from Ghidra
-        """
-        data = {
-            "binary": binary,
-            "project_name": project_name,
-            "script_file": script_file,
-            "analysis_timeout": analysis_timeout,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Ghidra analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/ghidra", data)
-        if result.get("success"):
-            logger.info(f"✅ Ghidra analysis completed for {binary}")
-        else:
-            logger.error(f"❌ Ghidra analysis failed for {binary}")
-        return result
-
-    @mcp.tool()
-    def pwntools_exploit(script_content: str = "", target_binary: str = "",
-                        target_host: str = "", target_port: int = 0,
-                        exploit_type: str = "local", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Pwntools for exploit development and automation.
-
-        Args:
-            script_content: Python script content using pwntools
-            target_binary: Local binary to exploit
-            target_host: Remote host to connect to
-            target_port: Remote port to connect to
-            exploit_type: Type of exploit (local, remote, format_string, rop)
-            additional_args: Additional arguments
-
-        Returns:
-            Exploit execution results
-        """
-        data = {
-            "script_content": script_content,
-            "target_binary": target_binary,
-            "target_host": target_host,
-            "target_port": target_port,
-            "exploit_type": exploit_type,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Pwntools exploit: {exploit_type}")
-        result = hexstrike_client.safe_post("api/tools/pwntools", data)
-        if result.get("success"):
-            logger.info(f"✅ Pwntools exploit completed")
-        else:
-            logger.error(f"❌ Pwntools exploit failed")
-        return result
-
-    @mcp.tool()
-    def one_gadget_search(libc_path: str, level: int = 1, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute one_gadget to find one-shot RCE gadgets in libc.
-
-        Args:
-            libc_path: Path to libc binary
-            level: Constraint level (0, 1, 2)
-            additional_args: Additional one_gadget arguments
-
-        Returns:
-            One-shot RCE gadget search results
-        """
-        data = {
-            "libc_path": libc_path,
-            "level": level,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting one_gadget analysis: {libc_path}")
-        result = hexstrike_client.safe_post("api/tools/one-gadget", data)
-        if result.get("success"):
-            logger.info(f"✅ one_gadget analysis completed")
-        else:
-            logger.error(f"❌ one_gadget analysis failed")
-        return result
-
-    @mcp.tool()
-    def libc_database_lookup(action: str = "find", symbols: str = "",
-                            libc_id: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute libc-database for libc identification and offset lookup.
-
-        Args:
-            action: Action to perform (find, dump, download)
-            symbols: Symbols with offsets for find action (format: "symbol1:offset1 symbol2:offset2")
-            libc_id: Libc ID for dump/download actions
-            additional_args: Additional arguments
-
-        Returns:
-            Libc database lookup results
-        """
-        data = {
-            "action": action,
-            "symbols": symbols,
-            "libc_id": libc_id,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting libc-database {action}: {symbols or libc_id}")
-        result = hexstrike_client.safe_post("api/tools/libc-database", data)
-        if result.get("success"):
-            logger.info(f"✅ libc-database {action} completed")
-        else:
-            logger.error(f"❌ libc-database {action} failed")
-        return result
-
-    @mcp.tool()
-    def gdb_peda_debug(binary: str = "", commands: str = "", attach_pid: int = 0,
-                      core_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute GDB with PEDA for enhanced debugging and exploitation.
-
-        Args:
-            binary: Binary to debug
-            commands: GDB commands to execute
-            attach_pid: Process ID to attach to
-            core_file: Core dump file to analyze
-            additional_args: Additional GDB arguments
-
-        Returns:
-            Enhanced debugging results with PEDA
-        """
-        data = {
-            "binary": binary,
-            "commands": commands,
-            "attach_pid": attach_pid,
-            "core_file": core_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting GDB-PEDA analysis: {binary or f'PID {attach_pid}' or core_file}")
-        result = hexstrike_client.safe_post("api/tools/gdb-peda", data)
-        if result.get("success"):
-            logger.info(f"✅ GDB-PEDA analysis completed")
-        else:
-            logger.error(f"❌ GDB-PEDA analysis failed")
-        return result
-
-    @mcp.tool()
-    def angr_symbolic_execution(binary: str, script_content: str = "",
-                               find_address: str = "", avoid_addresses: str = "",
-                               analysis_type: str = "symbolic", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute angr for symbolic execution and binary analysis.
-
-        Args:
-            binary: Binary to analyze
-            script_content: Custom angr script content
-            find_address: Address to find during symbolic execution
-            avoid_addresses: Comma-separated addresses to avoid
-            analysis_type: Type of analysis (symbolic, cfg, static)
-            additional_args: Additional arguments
-
-        Returns:
-            Symbolic execution and binary analysis results
-        """
-        data = {
-            "binary": binary,
-            "script_content": script_content,
-            "find_address": find_address,
-            "avoid_addresses": avoid_addresses,
-            "analysis_type": analysis_type,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting angr analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/angr", data)
-        if result.get("success"):
-            logger.info(f"✅ angr analysis completed")
-        else:
-            logger.error(f"❌ angr analysis failed")
-        return result
-
-    @mcp.tool()
-    def ropper_gadget_search(binary: str, gadget_type: str = "rop", quality: int = 1,
-                            arch: str = "", search_string: str = "",
-                            additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute ropper for advanced ROP/JOP gadget searching.
-
-        Args:
-            binary: Binary to search for gadgets
-            gadget_type: Type of gadgets (rop, jop, sys, all)
-            quality: Gadget quality level (1-5)
-            arch: Target architecture (x86, x86_64, arm, etc.)
-            search_string: Specific gadget pattern to search for
-            additional_args: Additional ropper arguments
-
-        Returns:
-            Advanced ROP/JOP gadget search results
-        """
-        data = {
-            "binary": binary,
-            "gadget_type": gadget_type,
-            "quality": quality,
-            "arch": arch,
-            "search_string": search_string,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting ropper analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/ropper", data)
-        if result.get("success"):
-            logger.info(f"✅ ropper analysis completed")
-        else:
-            logger.error(f"❌ ropper analysis failed")
-        return result
-
-    @mcp.tool()
-    def pwninit_setup(binary: str, libc: str = "", ld: str = "",
-                     template_type: str = "python", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute pwninit for CTF binary exploitation setup.
-
-        Args:
-            binary: Binary file to set up
-            libc: Libc file to use
-            ld: Loader file to use
-            template_type: Template type (python, c)
-            additional_args: Additional pwninit arguments
-
-        Returns:
-            CTF binary exploitation setup results
-        """
-        data = {
-            "binary": binary,
-            "libc": libc,
-            "ld": ld,
-            "template_type": template_type,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting pwninit setup: {binary}")
-        result = hexstrike_client.safe_post("api/tools/pwninit", data)
-        if result.get("success"):
-            logger.info(f"✅ pwninit setup completed")
-        else:
-            logger.error(f"❌ pwninit setup failed")
-        return result
-
-    @mcp.tool()
-    def feroxbuster_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", threads: int = 10, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Feroxbuster for recursive content discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            wordlist: Wordlist file to use
-            threads: Number of threads
-            additional_args: Additional Feroxbuster arguments
-
-        Returns:
-            Content discovery results
-        """
-        data = {
-            "url": url,
-            "wordlist": wordlist,
-            "threads": threads,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Feroxbuster scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/feroxbuster", data)
-        if result.get("success"):
-            logger.info(f"✅ Feroxbuster scan completed for {url}")
-        else:
-            logger.error(f"❌ Feroxbuster scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def dotdotpwn_scan(target: str, module: str = "http", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute DotDotPwn for directory traversal testing with enhanced logging.
-
-        Args:
-            target: The target hostname or IP
-            module: Module to use (http, ftp, tftp, etc.)
-            additional_args: Additional DotDotPwn arguments
-
-        Returns:
-            Directory traversal test results
-        """
-        data = {
-            "target": target,
-            "module": module,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting DotDotPwn scan: {target}")
-        result = hexstrike_client.safe_post("api/tools/dotdotpwn", data)
-        if result.get("success"):
-            logger.info(f"✅ DotDotPwn scan completed for {target}")
-        else:
-            logger.error(f"❌ DotDotPwn scan failed for {target}")
-        return result
-
-    @mcp.tool()
-    def xsser_scan(url: str, params: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute XSSer for XSS vulnerability testing with enhanced logging.
-
-        Args:
-            url: The target URL
-            params: Parameters to test
-            additional_args: Additional XSSer arguments
-
-        Returns:
-            XSS vulnerability test results
-        """
-        data = {
-            "url": url,
-            "params": params,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting XSSer scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/xsser", data)
-        if result.get("success"):
-            logger.info(f"✅ XSSer scan completed for {url}")
-        else:
-            logger.error(f"❌ XSSer scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def wfuzz_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Wfuzz for web application fuzzing with enhanced logging.
-
-        Args:
-            url: The target URL (use FUZZ where you want to inject payloads)
-            wordlist: Wordlist file to use
-            additional_args: Additional Wfuzz arguments
-
-        Returns:
-            Web application fuzzing results
-        """
-        data = {
-            "url": url,
-            "wordlist": wordlist,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Wfuzz scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/wfuzz", data)
-        if result.get("success"):
-            logger.info(f"✅ Wfuzz scan completed for {url}")
-        else:
-            logger.error(f"❌ Wfuzz scan failed for {url}")
-        return result
-
-    # ============================================================================
-    # ENHANCED WEB APPLICATION SECURITY TOOLS (v6.0)
-    # ============================================================================
-
-    @mcp.tool()
-    def dirsearch_scan(url: str, extensions: str = "php,html,js,txt,xml,json",
-                      wordlist: str = "/usr/share/wordlists/dirsearch/common.txt",
-                      threads: int = 30, recursive: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Dirsearch for advanced directory and file discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            extensions: File extensions to search for
-            wordlist: Wordlist file to use
-            threads: Number of threads to use
-            recursive: Enable recursive scanning
-            additional_args: Additional Dirsearch arguments
-
-        Returns:
-            Advanced directory discovery results
-        """
-        data = {
-            "url": url,
-            "extensions": extensions,
-            "wordlist": wordlist,
-            "threads": threads,
-            "recursive": recursive,
-            "additional_args": additional_args
-        }
-        logger.info(f"📁 Starting Dirsearch scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/dirsearch", data)
-        if result.get("success"):
-            logger.info(f"✅ Dirsearch scan completed for {url}")
-        else:
-            logger.error(f"❌ Dirsearch scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def katana_crawl(url: str, depth: int = 3, js_crawl: bool = True,
-                    form_extraction: bool = True, output_format: str = "json",
-                    additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Katana for next-generation crawling and spidering with enhanced logging.
-
-        Args:
-            url: The target URL to crawl
-            depth: Crawling depth
-            js_crawl: Enable JavaScript crawling
-            form_extraction: Enable form extraction
-            output_format: Output format (json, txt)
-            additional_args: Additional Katana arguments
-
-        Returns:
-            Advanced web crawling results with endpoints and forms
-        """
-        data = {
-            "url": url,
-            "depth": depth,
-            "js_crawl": js_crawl,
-            "form_extraction": form_extraction,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"⚔️  Starting Katana crawl: {url}")
-        result = hexstrike_client.safe_post("api/tools/katana", data)
-        if result.get("success"):
-            logger.info(f"✅ Katana crawl completed for {url}")
-        else:
-            logger.error(f"❌ Katana crawl failed for {url}")
-        return result
-
-    @mcp.tool()
-    def gau_discovery(domain: str, providers: str = "wayback,commoncrawl,otx,urlscan",
-                     include_subs: bool = True, blacklist: str = "png,jpg,gif,jpeg,swf,woff,svg,pdf,css,ico",
-                     additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Gau (Get All URLs) for URL discovery from multiple sources with enhanced logging.
-
-        Args:
-            domain: The target domain
-            providers: Data providers to use
-            include_subs: Include subdomains
-            blacklist: File extensions to blacklist
-            additional_args: Additional Gau arguments
-
-        Returns:
-            Comprehensive URL discovery results from multiple sources
-        """
-        data = {
-            "domain": domain,
-            "providers": providers,
-            "include_subs": include_subs,
-            "blacklist": blacklist,
-            "additional_args": additional_args
-        }
-        logger.info(f"📡 Starting Gau URL discovery: {domain}")
-        result = hexstrike_client.safe_post("api/tools/gau", data)
-        if result.get("success"):
-            logger.info(f"✅ Gau URL discovery completed for {domain}")
-        else:
-            logger.error(f"❌ Gau URL discovery failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def waybackurls_discovery(domain: str, get_versions: bool = False,
-                             no_subs: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Waybackurls for historical URL discovery with enhanced logging.
-
-        Args:
-            domain: The target domain
-            get_versions: Get all versions of URLs
-            no_subs: Don't include subdomains
-            additional_args: Additional Waybackurls arguments
-
-        Returns:
-            Historical URL discovery results from Wayback Machine
-        """
-        data = {
-            "domain": domain,
-            "get_versions": get_versions,
-            "no_subs": no_subs,
-            "additional_args": additional_args
-        }
-        logger.info(f"🕰️  Starting Waybackurls discovery: {domain}")
-        result = hexstrike_client.safe_post("api/tools/waybackurls", data)
-        if result.get("success"):
-            logger.info(f"✅ Waybackurls discovery completed for {domain}")
-        else:
-            logger.error(f"❌ Waybackurls discovery failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def arjun_parameter_discovery(url: str, method: str = "GET", wordlist: str = "",
-                                 delay: int = 0, threads: int = 25, stable: bool = False,
-                                 additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Arjun for HTTP parameter discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            method: HTTP method to use
-            wordlist: Custom wordlist file
-            delay: Delay between requests
-            threads: Number of threads
-            stable: Use stable mode
-            additional_args: Additional Arjun arguments
-
-        Returns:
-            HTTP parameter discovery results
-        """
-        data = {
-            "url": url,
-            "method": method,
-            "wordlist": wordlist,
-            "delay": delay,
-            "threads": threads,
-            "stable": stable,
-            "additional_args": additional_args
-        }
-        logger.info(f"🎯 Starting Arjun parameter discovery: {url}")
-        result = hexstrike_client.safe_post("api/tools/arjun", data)
-        if result.get("success"):
-            logger.info(f"✅ Arjun parameter discovery completed for {url}")
-        else:
-            logger.error(f"❌ Arjun parameter discovery failed for {url}")
-        return result
-
-    @mcp.tool()
-    def paramspider_mining(domain: str, level: int = 2,
-                          exclude: str = "png,jpg,gif,jpeg,swf,woff,svg,pdf,css,ico",
-                          output: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute ParamSpider for parameter mining from web archives with enhanced logging.
-
-        Args:
-            domain: The target domain
-            level: Mining level depth
-            exclude: File extensions to exclude
-            output: Output file path
-            additional_args: Additional ParamSpider arguments
-
-        Returns:
-            Parameter mining results from web archives
-        """
-        data = {
-            "domain": domain,
-            "level": level,
-            "exclude": exclude,
-            "output": output,
-            "additional_args": additional_args
-        }
-        logger.info(f"🕷️  Starting ParamSpider mining: {domain}")
-        result = hexstrike_client.safe_post("api/tools/paramspider", data)
-        if result.get("success"):
-            logger.info(f"✅ ParamSpider mining completed for {domain}")
-        else:
-            logger.error(f"❌ ParamSpider mining failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def x8_parameter_discovery(url: str, wordlist: str = "/usr/share/wordlists/x8/params.txt",
-                              method: str = "GET", body: str = "", headers: str = "",
-                              additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute x8 for hidden parameter discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            wordlist: Parameter wordlist
-            method: HTTP method
-            body: Request body
-            headers: Custom headers
-            additional_args: Additional x8 arguments
-
-        Returns:
-            Hidden parameter discovery results
-        """
-        data = {
-            "url": url,
-            "wordlist": wordlist,
-            "method": method,
-            "body": body,
-            "headers": headers,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting x8 parameter discovery: {url}")
-        result = hexstrike_client.safe_post("api/tools/x8", data)
-        if result.get("success"):
-            logger.info(f"✅ x8 parameter discovery completed for {url}")
-        else:
-            logger.error(f"❌ x8 parameter discovery failed for {url}")
-        return result
-
-    @mcp.tool()
-    def jaeles_vulnerability_scan(url: str, signatures: str = "", config: str = "",
-                                 threads: int = 20, timeout: int = 20,
-                                 additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Jaeles for advanced vulnerability scanning with custom signatures.
-
-        Args:
-            url: The target URL
-            signatures: Custom signature path
-            config: Configuration file
-            threads: Number of threads
-            timeout: Request timeout
-            additional_args: Additional Jaeles arguments
-
-        Returns:
-            Advanced vulnerability scanning results with custom signatures
-        """
-        data = {
-            "url": url,
-            "signatures": signatures,
-            "config": config,
-            "threads": threads,
-            "timeout": timeout,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔬 Starting Jaeles vulnerability scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/jaeles", data)
-        if result.get("success"):
-            logger.info(f"✅ Jaeles vulnerability scan completed for {url}")
-        else:
-            logger.error(f"❌ Jaeles vulnerability scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def dalfox_xss_scan(url: str, pipe_mode: bool = False, blind: bool = False,
-                       mining_dom: bool = True, mining_dict: bool = True,
-                       custom_payload: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Dalfox for advanced XSS vulnerability scanning with enhanced logging.
-
-        Args:
-            url: The target URL
-            pipe_mode: Use pipe mode for input
-            blind: Enable blind XSS testing
-            mining_dom: Enable DOM mining
-            mining_dict: Enable dictionary mining
-            custom_payload: Custom XSS payload
-            additional_args: Additional Dalfox arguments
-
-        Returns:
-            Advanced XSS vulnerability scanning results
-        """
-        data = {
-            "url": url,
-            "pipe_mode": pipe_mode,
-            "blind": blind,
-            "mining_dom": mining_dom,
-            "mining_dict": mining_dict,
-            "custom_payload": custom_payload,
-            "additional_args": additional_args
-        }
-        logger.info(f"🎯 Starting Dalfox XSS scan: {url if url else 'pipe mode'}")
-        result = hexstrike_client.safe_post("api/tools/dalfox", data)
-        if result.get("success"):
-            logger.info(f"✅ Dalfox XSS scan completed")
-        else:
-            logger.error(f"❌ Dalfox XSS scan failed")
-        return result
-
-    @mcp.tool()
-    def httpx_probe(target: str, probe: bool = True, tech_detect: bool = False,
-                   status_code: bool = False, content_length: bool = False,
-                   title: bool = False, web_server: bool = False, threads: int = 50,
-                   additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute httpx for fast HTTP probing and technology detection.
-
-        Args:
-            target: Target file or single URL
-            probe: Enable probing
-            tech_detect: Enable technology detection
-            status_code: Show status codes
-            content_length: Show content length
-            title: Show page titles
-            web_server: Show web server
-            threads: Number of threads
-            additional_args: Additional httpx arguments
-
-        Returns:
-            Fast HTTP probing results with technology detection
-        """
-        data = {
-            "target": target,
-            "probe": probe,
-            "tech_detect": tech_detect,
-            "status_code": status_code,
-            "content_length": content_length,
-            "title": title,
-            "web_server": web_server,
-            "threads": threads,
-            "additional_args": additional_args
-        }
-        logger.info(f"🌍 Starting httpx probe: {target}")
-        result = hexstrike_client.safe_post("api/tools/httpx", data)
-        if result.get("success"):
-            logger.info(f"✅ httpx probe completed for {target}")
-        else:
-            logger.error(f"❌ httpx probe failed for {target}")
-        return result
-
-    @mcp.tool()
-    def anew_data_processing(input_data: str, output_file: str = "",
-                            additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute anew for appending new lines to files (useful for data processing).
-
-        Args:
-            input_data: Input data to process
-            output_file: Output file path
-            additional_args: Additional anew arguments
-
-        Returns:
-            Data processing results with unique line filtering
-        """
-        data = {
-            "input_data": input_data,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info("📝 Starting anew data processing")
-        result = hexstrike_client.safe_post("api/tools/anew", data)
-        if result.get("success"):
-            logger.info("✅ anew data processing completed")
-        else:
-            logger.error("❌ anew data processing failed")
-        return result
-
-    @mcp.tool()
-    def qsreplace_parameter_replacement(urls: str, replacement: str = "FUZZ",
-                                       additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute qsreplace for query string parameter replacement.
-
-        Args:
-            urls: URLs to process
-            replacement: Replacement string for parameters
-            additional_args: Additional qsreplace arguments
-
-        Returns:
-            Parameter replacement results for fuzzing
-        """
-        data = {
-            "urls": urls,
-            "replacement": replacement,
-            "additional_args": additional_args
-        }
-        logger.info("🔄 Starting qsreplace parameter replacement")
-        result = hexstrike_client.safe_post("api/tools/qsreplace", data)
-        if result.get("success"):
-            logger.info("✅ qsreplace parameter replacement completed")
-        else:
-            logger.error("❌ qsreplace parameter replacement failed")
-        return result
-
-    @mcp.tool()
-    def uro_url_filtering(urls: str, whitelist: str = "", blacklist: str = "",
-                         additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute uro for filtering out similar URLs.
-
-        Args:
-            urls: URLs to filter
-            whitelist: Whitelist patterns
-            blacklist: Blacklist patterns
-            additional_args: Additional uro arguments
-
-        Returns:
-            Filtered URL results with duplicates removed
-        """
-        data = {
-            "urls": urls,
-            "whitelist": whitelist,
-            "blacklist": blacklist,
-            "additional_args": additional_args
-        }
-        logger.info("🔍 Starting uro URL filtering")
-        result = hexstrike_client.safe_post("api/tools/uro", data)
-        if result.get("success"):
-            logger.info("✅ uro URL filtering completed")
-        else:
-            logger.error("❌ uro URL filtering failed")
-        return result
-
+    # Register enhanced binary analysis and exploitation tools
+    from mcp_tools.enhanced_binary_analysis_and_exploitation import register_enhanced_binary_analysis_and_exploitation_tools
+    register_enhanced_binary_analysis_and_exploitation_tools(mcp, hexstrike_client, logger) 
+   
+    # Register enhanced web application security tools
+    from mcp_tools.enhanced_web_app_security import register_enhanced_web_app_security_tools
+    register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger)
+    
     # ============================================================================
     # AI-POWERED PAYLOAD GENERATION (v5.0 ENHANCEMENT)
     # ============================================================================
