+++ mcp_tools/enhanced_binary_analysis_and_exploitation.py
@@ -1,0 +1,363 @@
+# mcp_tools/enhanced_binary_analysis_and_exploitation.py
+
+from typing import Dict, Any
+
+def register_enhanced_binary_analysis_and_exploitation_tools(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def ghidra_analysis(binary: str, project_name: str = "hexstrike_analysis",
+                       script_file: str = "", analysis_timeout: int = 300,
+                       output_format: str = "xml", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Ghidra for advanced binary analysis and reverse engineering.
+
+        Args:
+            binary: Path to the binary file
+            project_name: Ghidra project name
+            script_file: Custom Ghidra script to run
+            analysis_timeout: Analysis timeout in seconds
+            output_format: Output format (xml, json)
+            additional_args: Additional Ghidra arguments
+
+        Returns:
+            Advanced binary analysis results from Ghidra
+        """
+        data = {
+            "binary": binary,
+            "project_name": project_name,
+            "script_file": script_file,
+            "analysis_timeout": analysis_timeout,
+            "output_format": output_format,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Ghidra analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/ghidra", data)
+        if result.get("success"):
+            logger.info(f"✅ Ghidra analysis completed for {binary}")
+        else:
+            logger.error(f"❌ Ghidra analysis failed for {binary}")
+        return result
+
+    @mcp.tool()
+    def pwntools_exploit(script_content: str = "", target_binary: str = "",
+                        target_host: str = "", target_port: int = 0,
+                        exploit_type: str = "local", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Pwntools for exploit development and automation.
+
+        Args:
+            script_content: Python script content using pwntools
+            target_binary: Local binary to exploit
+            target_host: Remote host to connect to
+            target_port: Remote port to connect to
+            exploit_type: Type of exploit (local, remote, format_string, rop)
+            additional_args: Additional arguments
+
+        Returns:
+            Exploit execution results
+        """
+        data = {
+            "script_content": script_content,
+            "target_binary": target_binary,
+            "target_host": target_host,
+            "target_port": target_port,
+            "exploit_type": exploit_type,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Pwntools exploit: {exploit_type}")
+        result = hexstrike_client.safe_post("api/tools/pwntools", data)
+        if result.get("success"):
+            logger.info(f"✅ Pwntools exploit completed")
+        else:
+            logger.error(f"❌ Pwntools exploit failed")
+        return result
+
+    @mcp.tool()
+    def one_gadget_search(libc_path: str, level: int = 1, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute one_gadget to find one-shot RCE gadgets in libc.
+
+        Args:
+            libc_path: Path to libc binary
+            level: Constraint level (0, 1, 2)
+            additional_args: Additional one_gadget arguments
+
+        Returns:
+            One-shot RCE gadget search results
+        """
+        data = {
+            "libc_path": libc_path,
+            "level": level,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting one_gadget analysis: {libc_path}")
+        result = hexstrike_client.safe_post("api/tools/one-gadget", data)
+        if result.get("success"):
+            logger.info(f"✅ one_gadget analysis completed")
+        else:
+            logger.error(f"❌ one_gadget analysis failed")
+        return result
+
+    @mcp.tool()
+    def libc_database_lookup(action: str = "find", symbols: str = "",
+                            libc_id: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute libc-database for libc identification and offset lookup.
+
+        Args:
+            action: Action to perform (find, dump, download)
+            symbols: Symbols with offsets for find action (format: "symbol1:offset1 symbol2:offset2")
+            libc_id: Libc ID for dump/download actions
+            additional_args: Additional arguments
+
+        Returns:
+            Libc database lookup results
+        """
+        data = {
+            "action": action,
+            "symbols": symbols,
+            "libc_id": libc_id,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting libc-database {action}: {symbols or libc_id}")
+        result = hexstrike_client.safe_post("api/tools/libc-database", data)
+        if result.get("success"):
+            logger.info(f"✅ libc-database {action} completed")
+        else:
+            logger.error(f"❌ libc-database {action} failed")
+        return result
+
+    @mcp.tool()
+    def gdb_peda_debug(binary: str = "", commands: str = "", attach_pid: int = 0,
+                      core_file: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute GDB with PEDA for enhanced debugging and exploitation.
+
+        Args:
+            binary: Binary to debug
+            commands: GDB commands to execute
+            attach_pid: Process ID to attach to
+            core_file: Core dump file to analyze
+            additional_args: Additional GDB arguments
+
+        Returns:
+            Enhanced debugging results with PEDA
+        """
+        data = {
+            "binary": binary,
+            "commands": commands,
+            "attach_pid": attach_pid,
+            "core_file": core_file,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting GDB-PEDA analysis: {binary or f'PID {attach_pid}' or core_file}")
+        result = hexstrike_client.safe_post("api/tools/gdb-peda", data)
+        if result.get("success"):
+            logger.info(f"✅ GDB-PEDA analysis completed")
+        else:
+            logger.error(f"❌ GDB-PEDA analysis failed")
+        return result
+
+    @mcp.tool()
+    def angr_symbolic_execution(binary: str, script_content: str = "",
+                               find_address: str = "", avoid_addresses: str = "",
+                               analysis_type: str = "symbolic", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute angr for symbolic execution and binary analysis.
+
+        Args:
+            binary: Binary to analyze
+            script_content: Custom angr script content
+            find_address: Address to find during symbolic execution
+            avoid_addresses: Comma-separated addresses to avoid
+            analysis_type: Type of analysis (symbolic, cfg, static)
+            additional_args: Additional arguments
+
+        Returns:
+            Symbolic execution and binary analysis results
+        """
+        data = {
+            "binary": binary,
+            "script_content": script_content,
+            "find_address": find_address,
+            "avoid_addresses": avoid_addresses,
+            "analysis_type": analysis_type,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting angr analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/angr", data)
+        if result.get("success"):
+            logger.info(f"✅ angr analysis completed")
+        else:
+            logger.error(f"❌ angr analysis failed")
+        return result
+
+    @mcp.tool()
+    def ropper_gadget_search(binary: str, gadget_type: str = "rop", quality: int = 1,
+                            arch: str = "", search_string: str = "",
+                            additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute ropper for advanced ROP/JOP gadget searching.
+
+        Args:
+            binary: Binary to search for gadgets
+            gadget_type: Type of gadgets (rop, jop, sys, all)
+            quality: Gadget quality level (1-5)
+            arch: Target architecture (x86, x86_64, arm, etc.)
+            search_string: Specific gadget pattern to search for
+            additional_args: Additional ropper arguments
+
+        Returns:
+            Advanced ROP/JOP gadget search results
+        """
+        data = {
+            "binary": binary,
+            "gadget_type": gadget_type,
+            "quality": quality,
+            "arch": arch,
+            "search_string": search_string,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting ropper analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/ropper", data)
+        if result.get("success"):
+            logger.info(f"✅ ropper analysis completed")
+        else:
+            logger.error(f"❌ ropper analysis failed")
+        return result
+
+    @mcp.tool()
+    def pwninit_setup(binary: str, libc: str = "", ld: str = "",
+                     template_type: str = "python", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute pwninit for CTF binary exploitation setup.
+
+        Args:
+            binary: Binary file to set up
+            libc: Libc file to use
+            ld: Loader file to use
+            template_type: Template type (python, c)
+            additional_args: Additional pwninit arguments
+
+        Returns:
+            CTF binary exploitation setup results
+        """
+        data = {
+            "binary": binary,
+            "libc": libc,
+            "ld": ld,
+            "template_type": template_type,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting pwninit setup: {binary}")
+        result = hexstrike_client.safe_post("api/tools/pwninit", data)
+        if result.get("success"):
+            logger.info(f"✅ pwninit setup completed")
+        else:
+            logger.error(f"❌ pwninit setup failed")
+        return result
+
+    @mcp.tool()
+    def feroxbuster_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", threads: int = 10, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Feroxbuster for recursive content discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            wordlist: Wordlist file to use
+            threads: Number of threads
+            additional_args: Additional Feroxbuster arguments
+
+        Returns:
+            Content discovery results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "threads": threads,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Feroxbuster scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/feroxbuster", data)
+        if result.get("success"):
+            logger.info(f"✅ Feroxbuster scan completed for {url}")
+        else:
+            logger.error(f"❌ Feroxbuster scan failed for {url}")
+        return result
+
+    @mcp.tool()
+    def dotdotpwn_scan(target: str, module: str = "http", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute DotDotPwn for directory traversal testing with enhanced logging.
+
+        Args:
+            target: The target hostname or IP
+            module: Module to use (http, ftp, tftp, etc.)
+            additional_args: Additional DotDotPwn arguments
+
+        Returns:
+            Directory traversal test results
+        """
+        data = {
+            "target": target,
+            "module": module,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting DotDotPwn scan: {target}")
+        result = hexstrike_client.safe_post("api/tools/dotdotpwn", data)
+        if result.get("success"):
+            logger.info(f"✅ DotDotPwn scan completed for {target}")
+        else:
+            logger.error(f"❌ DotDotPwn scan failed for {target}")
+        return result
+
+    @mcp.tool()
+    def xsser_scan(url: str, params: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute XSSer for XSS vulnerability testing with enhanced logging.
+
+        Args:
+            url: The target URL
+            params: Parameters to test
+            additional_args: Additional XSSer arguments
+
+        Returns:
+            XSS vulnerability test results
+        """
+        data = {
+            "url": url,
+            "params": params,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting XSSer scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/xsser", data)
+        if result.get("success"):
+            logger.info(f"✅ XSSer scan completed for {url}")
+        else:
+            logger.error(f"❌ XSSer scan failed for {url}")
+        return result
+
+    @mcp.tool()
+    def wfuzz_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Wfuzz for web application fuzzing with enhanced logging.
+
+        Args:
+            url: The target URL (use FUZZ where you want to inject payloads)
+            wordlist: Wordlist file to use
+            additional_args: Additional Wfuzz arguments
+
+        Returns:
+            Web application fuzzing results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Wfuzz scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/wfuzz", data)
+        if result.get("success"):
+            logger.info(f"✅ Wfuzz scan completed for {url}")
+        else:
+            logger.error(f"❌ Wfuzz scan failed for {url}")
+        return result
