+++ mcp_tools/enhanced_web_app_security.py
@@ -1,0 +1,429 @@
+# mcp_tools/enhanced_web_app_security.py
+
+from typing import Dict, Any
+
+def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def dirsearch_scan(url: str, extensions: str = "php,html,js,txt,xml,json",
+                      wordlist: str = "/usr/share/wordlists/dirsearch/common.txt",
+                      threads: int = 30, recursive: bool = False, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Dirsearch for advanced directory and file discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            extensions: File extensions to search for
+            wordlist: Wordlist file to use
+            threads: Number of threads to use
+            recursive: Enable recursive scanning
+            additional_args: Additional Dirsearch arguments
+
+        Returns:
+            Advanced directory discovery results
+        """
+        data = {
+            "url": url,
+            "extensions": extensions,
+            "wordlist": wordlist,
+            "threads": threads,
+            "recursive": recursive,
+            "additional_args": additional_args
+        }
+        logger.info(f"📁 Starting Dirsearch scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/dirsearch", data)
+        if result.get("success"):
+            logger.info(f"✅ Dirsearch scan completed for {url}")
+        else:
+            logger.error(f"❌ Dirsearch scan failed for {url}")
+        return result
+
+    @mcp.tool()
+    def katana_crawl(url: str, depth: int = 3, js_crawl: bool = True,
+                    form_extraction: bool = True, output_format: str = "json",
+                    additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Katana for next-generation crawling and spidering with enhanced logging.
+
+        Args:
+            url: The target URL to crawl
+            depth: Crawling depth
+            js_crawl: Enable JavaScript crawling
+            form_extraction: Enable form extraction
+            output_format: Output format (json, txt)
+            additional_args: Additional Katana arguments
+
+        Returns:
+            Advanced web crawling results with endpoints and forms
+        """
+        data = {
+            "url": url,
+            "depth": depth,
+            "js_crawl": js_crawl,
+            "form_extraction": form_extraction,
+            "output_format": output_format,
+            "additional_args": additional_args
+        }
+        logger.info(f"⚔️  Starting Katana crawl: {url}")
+        result = hexstrike_client.safe_post("api/tools/katana", data)
+        if result.get("success"):
+            logger.info(f"✅ Katana crawl completed for {url}")
+        else:
+            logger.error(f"❌ Katana crawl failed for {url}")
+        return result
+
+    @mcp.tool()
+    def gau_discovery(domain: str, providers: str = "wayback,commoncrawl,otx,urlscan",
+                     include_subs: bool = True, blacklist: str = "png,jpg,gif,jpeg,swf,woff,svg,pdf,css,ico",
+                     additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Gau (Get All URLs) for URL discovery from multiple sources with enhanced logging.
+
+        Args:
+            domain: The target domain
+            providers: Data providers to use
+            include_subs: Include subdomains
+            blacklist: File extensions to blacklist
+            additional_args: Additional Gau arguments
+
+        Returns:
+            Comprehensive URL discovery results from multiple sources
+        """
+        data = {
+            "domain": domain,
+            "providers": providers,
+            "include_subs": include_subs,
+            "blacklist": blacklist,
+            "additional_args": additional_args
+        }
+        logger.info(f"📡 Starting Gau URL discovery: {domain}")
+        result = hexstrike_client.safe_post("api/tools/gau", data)
+        if result.get("success"):
+            logger.info(f"✅ Gau URL discovery completed for {domain}")
+        else:
+            logger.error(f"❌ Gau URL discovery failed for {domain}")
+        return result
+
+    @mcp.tool()
+    def waybackurls_discovery(domain: str, get_versions: bool = False,
+                             no_subs: bool = False, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Waybackurls for historical URL discovery with enhanced logging.
+
+        Args:
+            domain: The target domain
+            get_versions: Get all versions of URLs
+            no_subs: Don't include subdomains
+            additional_args: Additional Waybackurls arguments
+
+        Returns:
+            Historical URL discovery results from Wayback Machine
+        """
+        data = {
+            "domain": domain,
+            "get_versions": get_versions,
+            "no_subs": no_subs,
+            "additional_args": additional_args
+        }
+        logger.info(f"🕰️  Starting Waybackurls discovery: {domain}")
+        result = hexstrike_client.safe_post("api/tools/waybackurls", data)
+        if result.get("success"):
+            logger.info(f"✅ Waybackurls discovery completed for {domain}")
+        else:
+            logger.error(f"❌ Waybackurls discovery failed for {domain}")
+        return result
+
+    @mcp.tool()
+    def arjun_parameter_discovery(url: str, method: str = "GET", wordlist: str = "",
+                                 delay: int = 0, threads: int = 25, stable: bool = False,
+                                 additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Arjun for HTTP parameter discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            method: HTTP method to use
+            wordlist: Custom wordlist file
+            delay: Delay between requests
+            threads: Number of threads
+            stable: Use stable mode
+            additional_args: Additional Arjun arguments
+
+        Returns:
+            HTTP parameter discovery results
+        """
+        data = {
+            "url": url,
+            "method": method,
+            "wordlist": wordlist,
+            "delay": delay,
+            "threads": threads,
+            "stable": stable,
+            "additional_args": additional_args
+        }
+        logger.info(f"🎯 Starting Arjun parameter discovery: {url}")
+        result = hexstrike_client.safe_post("api/tools/arjun", data)
+        if result.get("success"):
+            logger.info(f"✅ Arjun parameter discovery completed for {url}")
+        else:
+            logger.error(f"❌ Arjun parameter discovery failed for {url}")
+        return result
+
+    @mcp.tool()
+    def paramspider_mining(domain: str, level: int = 2,
+                          exclude: str = "png,jpg,gif,jpeg,swf,woff,svg,pdf,css,ico",
+                          output: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute ParamSpider for parameter mining from web archives with enhanced logging.
+
+        Args:
+            domain: The target domain
+            level: Mining level depth
+            exclude: File extensions to exclude
+            output: Output file path
+            additional_args: Additional ParamSpider arguments
+
+        Returns:
+            Parameter mining results from web archives
+        """
+        data = {
+            "domain": domain,
+            "level": level,
+            "exclude": exclude,
+            "output": output,
+            "additional_args": additional_args
+        }
+        logger.info(f"🕷️  Starting ParamSpider mining: {domain}")
+        result = hexstrike_client.safe_post("api/tools/paramspider", data)
+        if result.get("success"):
+            logger.info(f"✅ ParamSpider mining completed for {domain}")
+        else:
+            logger.error(f"❌ ParamSpider mining failed for {domain}")
+        return result
+
+    @mcp.tool()
+    def x8_parameter_discovery(url: str, wordlist: str = "/usr/share/wordlists/x8/params.txt",
+                              method: str = "GET", body: str = "", headers: str = "",
+                              additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute x8 for hidden parameter discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            wordlist: Parameter wordlist
+            method: HTTP method
+            body: Request body
+            headers: Custom headers
+            additional_args: Additional x8 arguments
+
+        Returns:
+            Hidden parameter discovery results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "method": method,
+            "body": body,
+            "headers": headers,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting x8 parameter discovery: {url}")
+        result = hexstrike_client.safe_post("api/tools/x8", data)
+        if result.get("success"):
+            logger.info(f"✅ x8 parameter discovery completed for {url}")
+        else:
+            logger.error(f"❌ x8 parameter discovery failed for {url}")
+        return result
+
+    @mcp.tool()
+    def jaeles_vulnerability_scan(url: str, signatures: str = "", config: str = "",
+                                 threads: int = 20, timeout: int = 20,
+                                 additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Jaeles for advanced vulnerability scanning with custom signatures.
+
+        Args:
+            url: The target URL
+            signatures: Custom signature path
+            config: Configuration file
+            threads: Number of threads
+            timeout: Request timeout
+            additional_args: Additional Jaeles arguments
+
+        Returns:
+            Advanced vulnerability scanning results with custom signatures
+        """
+        data = {
+            "url": url,
+            "signatures": signatures,
+            "config": config,
+            "threads": threads,
+            "timeout": timeout,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔬 Starting Jaeles vulnerability scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/jaeles", data)
+        if result.get("success"):
+            logger.info(f"✅ Jaeles vulnerability scan completed for {url}")
+        else:
+            logger.error(f"❌ Jaeles vulnerability scan failed for {url}")
+        return result
+
+    @mcp.tool()
+    def dalfox_xss_scan(url: str, pipe_mode: bool = False, blind: bool = False,
+                       mining_dom: bool = True, mining_dict: bool = True,
+                       custom_payload: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Dalfox for advanced XSS vulnerability scanning with enhanced logging.
+
+        Args:
+            url: The target URL
+            pipe_mode: Use pipe mode for input
+            blind: Enable blind XSS testing
+            mining_dom: Enable DOM mining
+            mining_dict: Enable dictionary mining
+            custom_payload: Custom XSS payload
+            additional_args: Additional Dalfox arguments
+
+        Returns:
+            Advanced XSS vulnerability scanning results
+        """
+        data = {
+            "url": url,
+            "pipe_mode": pipe_mode,
+            "blind": blind,
+            "mining_dom": mining_dom,
+            "mining_dict": mining_dict,
+            "custom_payload": custom_payload,
+            "additional_args": additional_args
+        }
+        logger.info(f"🎯 Starting Dalfox XSS scan: {url if url else 'pipe mode'}")
+        result = hexstrike_client.safe_post("api/tools/dalfox", data)
+        if result.get("success"):
+            logger.info(f"✅ Dalfox XSS scan completed")
+        else:
+            logger.error(f"❌ Dalfox XSS scan failed")
+        return result
+
+    @mcp.tool()
+    def httpx_probe(target: str, probe: bool = True, tech_detect: bool = False,
+                   status_code: bool = False, content_length: bool = False,
+                   title: bool = False, web_server: bool = False, threads: int = 50,
+                   additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute httpx for fast HTTP probing and technology detection.
+
+        Args:
+            target: Target file or single URL
+            probe: Enable probing
+            tech_detect: Enable technology detection
+            status_code: Show status codes
+            content_length: Show content length
+            title: Show page titles
+            web_server: Show web server
+            threads: Number of threads
+            additional_args: Additional httpx arguments
+
+        Returns:
+            Fast HTTP probing results with technology detection
+        """
+        data = {
+            "target": target,
+            "probe": probe,
+            "tech_detect": tech_detect,
+            "status_code": status_code,
+            "content_length": content_length,
+            "title": title,
+            "web_server": web_server,
+            "threads": threads,
+            "additional_args": additional_args
+        }
+        logger.info(f"🌍 Starting httpx probe: {target}")
+        result = hexstrike_client.safe_post("api/tools/httpx", data)
+        if result.get("success"):
+            logger.info(f"✅ httpx probe completed for {target}")
+        else:
+            logger.error(f"❌ httpx probe failed for {target}")
+        return result
+
+    @mcp.tool()
+    def anew_data_processing(input_data: str, output_file: str = "",
+                            additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute anew for appending new lines to files (useful for data processing).
+
+        Args:
+            input_data: Input data to process
+            output_file: Output file path
+            additional_args: Additional anew arguments
+
+        Returns:
+            Data processing results with unique line filtering
+        """
+        data = {
+            "input_data": input_data,
+            "output_file": output_file,
+            "additional_args": additional_args
+        }
+        logger.info("📝 Starting anew data processing")
+        result = hexstrike_client.safe_post("api/tools/anew", data)
+        if result.get("success"):
+            logger.info("✅ anew data processing completed")
+        else:
+            logger.error("❌ anew data processing failed")
+        return result
+
+    @mcp.tool()
+    def qsreplace_parameter_replacement(urls: str, replacement: str = "FUZZ",
+                                       additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute qsreplace for query string parameter replacement.
+
+        Args:
+            urls: URLs to process
+            replacement: Replacement string for parameters
+            additional_args: Additional qsreplace arguments
+
+        Returns:
+            Parameter replacement results for fuzzing
+        """
+        data = {
+            "urls": urls,
+            "replacement": replacement,
+            "additional_args": additional_args
+        }
+        logger.info("🔄 Starting qsreplace parameter replacement")
+        result = hexstrike_client.safe_post("api/tools/qsreplace", data)
+        if result.get("success"):
+            logger.info("✅ qsreplace parameter replacement completed")
+        else:
+            logger.error("❌ qsreplace parameter replacement failed")
+        return result
+
+    @mcp.tool()
+    def uro_url_filtering(urls: str, whitelist: str = "", blacklist: str = "",
+                         additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute uro for filtering out similar URLs.
+
+        Args:
+            urls: URLs to filter
+            whitelist: Whitelist patterns
+            blacklist: Blacklist patterns
+            additional_args: Additional uro arguments
+
+        Returns:
+            Filtered URL results with duplicates removed
+        """
+        data = {
+            "urls": urls,
+            "whitelist": whitelist,
+            "blacklist": blacklist,
+            "additional_args": additional_args
+        }
+        logger.info("🔍 Starting uro URL filtering")
+        result = hexstrike_client.safe_post("api/tools/uro", data)
+        if result.get("success"):
+            logger.info("✅ uro URL filtering completed")
+        else:
+            logger.error("❌ uro URL filtering failed")
+        return result
